backend/Message_Passing_Module/query_injector/requirements.txt
functions-framework==3.*
google-cloud-pubsub>=2.22.0
